8주차 과제 재제출

+++ dayeon/likelion-assignment-crud/src/main/java/org/likelion/likelionassignmentcrud/order/api/dto/response/OrderInfoResDto.java
@@ -6,7 +6,7 @@ import org.likelion.likelionassignmentcrud.order.domain.Order;
 
 @Builder
 public record OrderInfoResDto(
-        Long date,
+        int date,
         Item item,
         int quantity,
         String member
